+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.0.5] - 2026-01-16
+### Changed
+- Improves canvas rendering performance with viewport culling
+- Enhances connection routing for nodes above source
+
+### Fixed
+- Ensures final line segment is long enough for arrows
+
 ## [0.0.4] - 2026-01-16
 ### Added
 - Join node type and custom routing
@@ -16,4 +24,5 @@ All notable changes to this project will be documented in this file.
 - Initial tagged release
 
 [0.0.4]: https://github.com/taurgis/SFCC-Manifold/compare/v0.0.3...v0.0.4
+[0.0.5]: https://github.com/taurgis/SFCC-Manifold/compare/v0.0.4...v0.0.5
 [0.0.3]: https://github.com/taurgis/SFCC-Manifold/releases/tag/v0.0.3